test(event-handler): split event-bulk-deploy.test.ts by scenario
Issue #1233 — 856-line / 115-expect monolith with assertion roulette across
30 cases. Split into 5 sibling files (each under 250 lines) keyed by scenario,
with shared fixtures (buildShared / sampleEvent / sampleTeams / NOW_MS)
extracted into a `.test-helpers.ts` (NOT `.test.ts`) module:

- event-bulk-deploy-happy.test.ts          (99 lines, 20 expects)
- event-bulk-deploy-errors.test.ts        (131 lines, 19 expects)
- event-bulk-deploy-idempotency.test.ts   (243 lines, 29 expects)
- event-bulk-deploy-persistence.test.ts   (125 lines, 14 expects)
- event-bulk-deploy-verification.test.ts  (177 lines, 33 expects)
- event-bulk-deploy.test-helpers.ts       (119 lines, shared fixtures)

Total 115 expects preserved across the split. All 30 tests pass via the same
`bulk-deploy.ts` façade (no production code touched).

Closes #1233
